Move Tools.java to utility package
