+++ src/main/java/com/alphatica/genotick/genotick/SimpleEngine.java
@@ -15,6 +15,7 @@ import com.alphatica.genotick.timepoint.TimePointExecutor;
 import com.alphatica.genotick.timepoint.TimePointResult;
 import com.alphatica.genotick.ui.UserInputOutputFactory;
 import com.alphatica.genotick.ui.UserOutput;
+import com.alphatica.genotick.utility.Tools;
 
 import static com.alphatica.genotick.utility.Assert.gassert;
